+++ src/main/php/lang/reflection/Constructor.class.php
@@ -1,6 +1,6 @@
 <?php namespace lang\reflection;
 
-use lang\{Reflection, Type, IllegalArgumentException};
+use lang\Reflection;
 
 class Constructor extends Routine {
   private $class;
@@ -43,12 +43,26 @@ class Constructor extends Routine {
    * Creates a new instance of the type this constructor belongs to
    *
    * @param  var[] $args
+   * @param  ?string|?lang.XPClass|?lang.reflection.Type $context
    * @return object
    * @throws lang.reflection.InvocationFailed
    * @throws lang.reflection.CannotInstantiate
    */
-  public function newInstance($args= []) {
+  public function newInstance($args= [], $context= null) {
     try {
+
+      // Workaround for non-public constructors: Set accessible, then manually
+      // invoke after creating an instance without invoking the constructor.
+      if ($context && !$this->reflect->isPublic()) {
+        $t= $context instanceof Type ? $context : Reflection::of($context);
+        if ($t->is($this->class->name)) {
+          $instance= $this->class->newInstanceWithoutConstructor();
+          $this->reflect->setAccessible(true);
+          $this->reflect->invokeArgs($instance, $args);
+          return $instance;
+        }
+      }
+
       return $this->class->newInstanceArgs($args);
     } catch (\ReflectionException $e) {
       throw new CannotInstantiate($this->class->name, $e);